Implemented Workout tables, Added input selection dropdown for workouts, Lifts associated with workout ids
Note: On edit, didn't allow option for workout id to be edited but had to send it over as a value in the packaged data that was PUT so the JSON format could be interpreted properly. Probably don't want this value to be edited in the future either since users will just add or remove lifts within their associated workout card (or move to a different workout card) rather than having an option to change the workout id for each lift.